Ora il carburante è visualizzato con un float

+++ Project/Project/Nave.cpp
@@ -58,7 +58,7 @@ void Nave::spara()
 	}
 }
 
-int Nave::getCarburante() {
+float Nave::getCarburante() {
 	return carburante_;
 }
